Final Code.cpp
all of the algorithm has been implemented however, the gsm module is not sending sms. It can be due to the lack of power from the voltage but we already have a barrel jack with 12V, another possible thing that this is not working is that the signal is poor or cannot use the 2g from the simcard.